backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Object Storage
minio==7.2.0

# Message Queue
aio-pika==9.3.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Processing
numpy==1.25.2
pandas==2.1.4

# Utilities
python-dateutil==2.8.2
typing-extensions==4.8.0

# Logging
loguru==0.7.2

# Validation
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Background Tasks
celery==5.3.4

# Monitoring
prometheus-client==0.19.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools (optional)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1